deforestation/requirements.txt
geopandas~=1.0.0
numpy~=1.26.4
matplotlib~=3.9.0
pandas~=2.2.2
rasterio~=1.4.3
shapely~=2.0.3
boto3~=1.34
botocore
tqdm~=4.66
scikit-learn~=1.4.0
pyarrow~=16.0.0
pyyaml~=6.0
xgboost~=2.0.0
joblib~=1.4.0
torch
cleanlab~=2.6.0
ruptures~=1.1.9
